+++ Shared/View/Components/WatchlistItem/WatchlistItemPoster.swift
@@ -28,7 +28,7 @@ struct WatchlistItemPoster: View {
         .accessibilityLabel(Text(content.itemTitle))
         .sheet(isPresented: $showNote) {
             NavigationStack {
-                WatchlistItemNoteView(item: content, showView: $showNote)
+                WatchlistItemNoteView(id: content.notificationID, showView: $showNote)
             }
             .presentationDetents([.medium, .large])
 #if os(macOS)
