+++ Shared/View/WatchlistItemNoteView.swift
@@ -8,42 +8,50 @@
 import SwiftUI
 
 struct WatchlistItemNoteView: View {
-    let item: WatchlistItem
+    let id: String
+    @State var item: WatchlistItem?
     @Binding var showView: Bool
     @State private var note = String()
     @State private var rating = 0
+    @State private var isLoading = true
+    let persistence = PersistenceController.shared
     var body: some View {
         Form {
-            Text("reviewOf \(item.itemTitle)")
-            Section {
-                CenterHorizontalView {
-                    RatingView(rating: $rating)
-                }
-            } header: {
-                Text("Rating") 
+            if isLoading {
+                ProgressView()
+            } else {
+                if let item {
+                    Text("reviewOf \(item.itemTitle)")
+                    Section {
+                        CenterHorizontalView {
+                            RatingView(rating: $rating)
+                        }
+                    } header: {
+                        Text("Rating")
 #if os(macOS)
-                    .foregroundColor(.secondary)
-                    .font(.callout)
+                            .foregroundColor(.secondary)
+                            .font(.callout)
 #endif
-            }
-            Section {
+                    }
+                    Section {
 #if os(iOS) || os(macOS)
-                TextEditor(text: $note)
-                    .frame(minHeight: 150)
+                        TextEditor(text: $note)
+                            .frame(minHeight: 150)
 #endif
-            } header: {
-                Text("Notes")
+                    } header: {
+                        Text("Notes")
 #if os(macOS)
-                    .foregroundColor(.secondary)
-                    .font(.callout)
+                            .foregroundColor(.secondary)
+                            .font(.callout)
 #endif
+                    }
+                } else {
+                    ProgressView()
+                }
             }
         }
         .navigationTitle("reviewTitle")
-        .onAppear {
-            if note.isEmpty { note = item.userNotes }
-            rating = Int(item.userRating)
-        }
+        .onAppear { load() }
         .toolbar {
 #if os(iOS)
             ToolbarItem(placement: .navigationBarLeading) { doneButton }
@@ -58,6 +66,17 @@ struct WatchlistItemNoteView: View {
 #endif
     }
     
+    private func load() {
+        let item = try? persistence.fetch(for: id)
+        guard let item else {
+            return
+        }
+        if note.isEmpty { note = item.userNotes }
+        rating = Int(item.userRating)
+        self.item = item
+        withAnimation { isLoading = false }
+    }
+    
     private var doneButton: some View {
         Button("Cancel", action: dismiss)
     }
@@ -67,19 +86,18 @@ struct WatchlistItemNoteView: View {
     }
     
     private func save() {
-        PersistenceController.shared.updateReview(for: item, rating: rating, notes: note)
+        guard let item else { return }
+        persistence.updateReview(for: item, rating: rating, notes: note)
         dismiss()
     }
     
-    private func dismiss() {
-        showView.toggle()
-    }
+    private func dismiss() { showView.toggle() }
 }
 
 struct WatchlistItemNoteView_Previews: PreviewProvider {
     static var previews: some View {
         NavigationStack {
-            WatchlistItemNoteView(item: .example, showView: .constant(true))
+            WatchlistItemNoteView(id: ItemContent.previewContent.itemNotificationID, showView: .constant(true))
         }
     }
 }
